+++ nodes/opcodes.js
@@ -19,6 +19,7 @@ module.exports = {
   SetTimeProfile: 0x88,
   ClearTimeProfiles: 0x8a,
   SetPCControl: 0xa0,
+  SetInterlock: 0xa2,
   ClearTaskList: 0xa6,
   AddTask: 0xa8,
   RefreshTaskList: 0xac,