+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+1. `uhppoted-set-interlock` node.
+
 
 ## [1.1.3](https://github.com/uhppoted/node-red-contrib-uhppoted/releases/tag/v1.1.3) - 2023-03-17